+++ Lemegeton/Content/Timelines.cs
@@ -300,7 +300,7 @@ namespace Lemegeton.Content
                 }
             }
 
-            private void _state_OnZoneChange(ushort newZone)            
+            private void _state_OnZoneChange(uint newZone)            
             {
                 if (recState != RecordingStateEnum.Idle)
                 {
